HBASE-1916 FindBugs and javadoc warnings cleanup
git-svn-id: https://svn.apache.org/repos/asf/hadoop/hbase/trunk@826569 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/org/apache/hadoop/hbase/MapFilePerformanceEvaluation.java
@@ -330,7 +330,8 @@ public class MapFilePerformanceEvaluation {
     }
     
     private ImmutableBytesWritable getGaussianRandomRow() {
-      int r = (int) randomData.nextGaussian(totalRows / 2, totalRows / 10);
+      int r = (int) randomData.nextGaussian((double)totalRows / 2.0,
+          (double)totalRows / 10.0);
       return format(r, key);
     }
